+++ src/main/java/minegame159/aoc2023/days/Day02.java
@@ -12,7 +12,7 @@ public class Day02 extends Day {
     }
 
     @Override
-    public int calculateFirst(List<String> input) {
+    public long calculateFirst(List<String> input) {
         int sum = 0;
 
         for (String line : input) {
@@ -24,7 +24,7 @@ public class Day02 extends Day {
     }
 
     @Override
-    public int calculateSecond(List<String> input) {
+    public long calculateSecond(List<String> input) {
         int sum = 0;
 
         for (String line : input) {
