+++ src/main/java/minegame159/aoc2023/days/Day04.java
@@ -11,7 +11,7 @@ public class Day04 extends Day {
     }
 
     @Override
-    public int calculateFirst(List<String> input) {
+    public long calculateFirst(List<String> input) {
         int sum = 0;
 
         for (String line : input) {
@@ -23,7 +23,7 @@ public class Day04 extends Day {
     }
 
     @Override
-    public int calculateSecond(List<String> input) {
+    public long calculateSecond(List<String> input) {
         int[] cards = new int[input.size()];
         int total = input.size();
